Revert "[DF-5671] Fixed cast warnings"
This reverts commit 8fcd686c89455b4bc6d00332d107310b2d58eafe [formerly 54f4d56e4ffae51bfe4bca0996ef6a200a0ac375].


Former-commit-id: 39ba6e600776e6cfb7da64fa4faea6f46cb7ad69

+++ Sources/Internal/Scene3D/Entity.cpp
@@ -118,7 +118,7 @@ void Entity::AddComponent(Component * component)
 {
 	component->SetEntity(this);
 		
-	uint64 componentType = component->GetType();
+	uint32 componentType = component->GetType();
 	if(USE_VECTOR(componentType))
 	{
 		DVASSERT(NULL == components[componentType]);
@@ -153,7 +153,7 @@ void Entity::AddComponent(Component * component)
 #endif
 	}
 		
-	componentFlags |= (1 << (uint64)component->GetType());
+	componentFlags |= 1 << component->GetType();
 
 	if (scene)
 		scene->RegisterComponent(this, component);
@@ -198,7 +198,7 @@ void Entity::RemoveAllComponents()
 void Entity::RemoveComponent(Component * component)
 {
 	int componentCount = 0;
-	uint64 componentType = component->GetType();
+	uint32 componentType = component->GetType();
     
     DetachComponent(component);
 
@@ -228,7 +228,7 @@ void Entity::DetachComponent( Component * component )
     if ( scene )
         scene->UnregisterComponent( this, component );
 
-    uint64 componentType = component->GetType();
+    uint32 componentType = component->GetType();
     uint32 componentCount = 0;
 
     if (USE_VECTOR(componentType))
@@ -287,7 +287,7 @@ void Entity::RemoveComponent(uint32 componentType, uint32 index)
 
 	Component* component = NULL;
 	int componentCount = 0;
-	if(USE_VECTOR((uint64)componentType))
+	if(USE_VECTOR(componentType))
 	{
 		component = components[componentType];
 		components[componentType] = NULL;
@@ -341,7 +341,7 @@ inline void Entity::CleanupComponent(Component* component, uint32 componentCount
 Component * Entity::GetComponent(uint32 componentType, uint32 index) const
 {
 	Component* component = NULL;
-	if(USE_VECTOR((uint64)componentType))
+	if(USE_VECTOR(componentType))
 	{
 		component = components[componentType];
 	}
@@ -428,7 +428,7 @@ uint32 Entity::GetComponentCount(uint32 componentType)
 {
 	int componentCount = 0;
 		
-	if(USE_VECTOR((uint64)componentType))
+	if(USE_VECTOR(componentType))
 	{
 		if(components[componentType] != NULL)
 		{
